added code to write one non-redundant sequences to file for orthoMCL analysis
For each Corset cluster, this will write the sequence with the longest blast hit from
an all-by-all blast search to a file for orthMCL analysis.

I've also moved the orthoMCL code to after this command, but I need to run the all-by-all
blast against all sequences prior to this step so I can determine which sequence to use
as the representative sequence.

It's probably a lot faster to filter the representative sequences out of all the blast hits
rather than redoing the blast, tho it's a bit tricky because I need to filter out results
where either the query OR the subject is not in the nr set.

This, in turn, is most easily accomplished by recording which sequences are redundant in
the db (as a separate step instead of what I'm doing now).