added code to plot histograms of lengths of coding portions of non-redundant sequences
The plots aren't perfect at this point. I would like to figure out how to write the
contig number for each one on the plots, as well as writing the species names inside the
plot with the same colour coding as the plots. Ideally, I would also plot the histogram
for the reference strain behind, but that may be more trouble than it's worth