Added verbose mode
The -v flag will now cause SQL commands to be printed to STDERR (for functions where
it has been implemented)

I'm tired of forgetting to update the print statements when I update the cur.execute
commands, so I refactored things so that a single command is used for both.

This would be easy if I used string formatting for the cur.execute command, but you're
not supposed to do that.

This means that I had to write a PrintCommand function which converts a command and a list
of options to a string for printing.

This means that the command and options have to be constructed separately, but that may
not be such a bad thing anyway.

I also wrote it so that options are wrapped in quotes when printed so that the statement
can be simply copied into mysql and should give the same result.